+++ services/src/ops/ops.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { OpsService } from './ops.service';
+import { OpsResolver } from './ops.resolver';
+
+@Module({
+  providers: [OpsService, OpsResolver],
+})
+export class OpsModule {}